+++ src/main/java/org/websoso/WSSServer/application/AuthApplication.java
@@ -17,7 +17,7 @@ import org.websoso.WSSServer.oauth2.service.KakaoService;
 import org.websoso.WSSServer.repository.RefreshTokenRepository;
 import org.websoso.WSSServer.oauth2.domain.RefreshToken;
 import org.websoso.WSSServer.user.domain.User;
-import org.websoso.WSSServer.oauth2.repository.UserDeviceRepository;
+import org.websoso.WSSServer.notification.repository.UserDeviceRepository;
 import org.websoso.WSSServer.user.service.UserService;
 
 @Service
