+++ src/main/java/org/websoso/WSSServer/notification/repository/UserDeviceRepository.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.oauth2.repository;
+package org.websoso.WSSServer.notification.repository;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
